+++ src/Ouzo/Core/Logger/StdOutputLogger.php
@@ -6,33 +6,16 @@
 namespace Ouzo\Logger;
 
 use Ouzo\Config;
-use Ouzo\Utilities\Arrays;
 use Ouzo\Utilities\Clock;
 
-class StdOutputLogger implements LoggerInterface
+class StdOutputLogger extends AbstractLogger
 {
-    private $_name;
-    /**
-     * @var MessageFormatter
-     */
-    private $_messageFormatter;
     private $_outputStreamIdentifier;
 
     public function __construct($name, $configuration, $outputStreamIdentifier = 'php')
     {
-        $this->_name = $name;
-        $messageFormatterClass = 'Ouzo\Logger\DefaultMessageFormatter';
-        $logger = Config::getValue('logger', $configuration);
-        if ($logger) {
-            $messageFormatterClass = Arrays::getValue($logger, 'formatter', $messageFormatterClass);
-        }
-        $this->_messageFormatter = new $messageFormatterClass();
+        parent::__construct($name, $configuration);
         $this->_outputStreamIdentifier = $outputStreamIdentifier;
-    }
-
-    public function setName($name)
-    {
-        $this->_name = $name;
     }
 
     private function _errorStreamName()
@@ -45,47 +28,40 @@ class StdOutputLogger implements LoggerInterface
         return $this->_outputStreamIdentifier . "://stdout";
     }
 
-    private function log($stdOut, $levelName, $message, $params)
+    private function _log($stdOut, $level, $levelName, $message, $params)
     {
-        $message = $this->_messageFormatter->format($this->_name, $levelName, $message);
-        if (!empty($params)) {
-            $message = call_user_func_array('sprintf', array_merge(array($message), $params));
-        }
-        $date = Clock::nowAsString();
-        $fileHandle = fopen($stdOut, 'a');
-        fwrite($fileHandle, "$date: $message\n");
-        fclose($fileHandle);
+        $this->log(function ($message) use ($stdOut) {
+            $date = Clock::nowAsString();
+            $fileHandle = fopen($stdOut, 'a');
+            fwrite($fileHandle, "$date: $message\n");
+            fclose($fileHandle);
+        }, $level, $levelName, $message, $params);
     }
 
     public function error($message, $params = null)
     {
-        $this->log($this->_errorStreamName(), 'Error', $message, $params);
+        $this->_log($this->_errorStreamName(), LOG_ERR, 'Error', $message, $params);
     }
 
     public function info($message, $params = null)
     {
-        $this->log($this->_standardStreamName(), 'Info', $message, $params);
+        $this->_log($this->_standardStreamName(), LOG_INFO, 'Info', $message, $params);
     }
 
     public function debug($message, $params = null)
     {
-        if ($this->_isDebug()) {
-            $this->log($this->_standardStreamName(), 'Debug', $message, $params);
+        if ($this->isDebug()) {
+            $this->_log($this->_standardStreamName(), LOG_DEBUG, 'Debug', $message, $params);
         }
-    }
-
-    private function _isDebug()
-    {
-        return Config::getValue('debug');
     }
 
     public function warning($message, $params = null)
     {
-        $this->log($this->_standardStreamName(), 'Warning', $message, $params);
+        $this->_log($this->_standardStreamName(), LOG_WARNING, 'Warning', $message, $params);
     }
 
     public function fatal($message, $params = null)
     {
-        $this->log($this->_errorStreamName(), 'Fatal', $message, $params);
+        $this->_log($this->_errorStreamName(), LOG_CRIT, 'Fatal', $message, $params);
     }
 }