+++ src/Ouzo/Core/Logger/SyslogLogger.php
@@ -6,17 +6,13 @@
 namespace Ouzo\Logger;
 
 use Ouzo\Config;
-use Ouzo\Utilities\Arrays;
 
-class SyslogLogger implements LoggerInterface
+class SyslogLogger extends AbstractLogger
 {
-    private $_name;
-    private $_configuration;
 
     public function __construct($name, $configuration)
     {
-        $this->_name = $name;
-        $this->_configuration = $configuration;
+        parent::__construct($name, $configuration);
     }
 
     public function __destruct()
@@ -24,58 +20,41 @@ class SyslogLogger implements LoggerInterface
         closelog();
     }
 
-    private function log($level, $levelName, $message, $params)
+    private function _log($level, $levelName, $message, $params)
     {
-        $messageFormatterClass = '\Ouzo\Logger\DefaultMessageFormatter';
-        $logger = Config::getValue('logger', $this->_configuration);
-        if ($logger) {
-            openlog($logger['ident'], $logger['option'], $logger['facility']);
-            $messageFormatterClass = Arrays::getValue($logger, 'formatter', $messageFormatterClass);
-        }
-
-        $messageFormatter = new $messageFormatterClass();
-        $message = $messageFormatter->format($this->_name, $levelName, $message);
-        if (!empty($params)) {
-            $message = call_user_func_array('sprintf', array_merge(array($message), $params));
-        }
-
-        syslog($level, $message);
+        $logger = $this->getLogger();
+        $this->log(function ($message) use ($logger, $level) {
+            if ($logger) {
+                openlog($logger['ident'], $logger['option'], $logger['facility']);
+            }
+            syslog($level, $message);
+        }, $level, $levelName, $message, $params);
     }
 
     public function error($message, $params = null)
     {
-        $this->log(LOG_ERR, 'Error', $message, $params);
+        $this->_log(LOG_ERR, 'Error', $message, $params);
     }
 
     public function info($message, $params = null)
     {
-        $this->log(LOG_INFO, 'Info', $message, $params);
+        $this->_log(LOG_INFO, 'Info', $message, $params);
     }
 
     public function debug($message, $params = null)
     {
-        if ($this->_isDebug()) {
-            $this->log(LOG_DEBUG, 'Debug', $message, $params);
+        if ($this->isDebug()) {
+            $this->_log(LOG_DEBUG, 'Debug', $message, $params);
         }
     }
 
     public function warning($message, $params = null)
     {
-        $this->log(LOG_WARNING, 'Warning', $message, $params);
+        $this->_log(LOG_WARNING, 'Warning', $message, $params);
     }
 
     public function fatal($message, $params = null)
     {
-        $this->log(LOG_CRIT, 'Fatal', $message, $params);
-    }
-
-    private function _isDebug()
-    {
-        return Config::getValue('debug');
-    }
-
-    public function setName($name)
-    {
-        $this->_name = $name;
+        $this->_log(LOG_CRIT, 'Fatal', $message, $params);
     }
 }